Project  VioletRed/requirements.txt
torch
transformers
tokenizers
pyyaml
numpy
accelerate